+++ Assets/Scripts/Data/Stats/Statistic.cs
@@ -2,11 +2,11 @@ using Data.Types;
 
 namespace Data.Stats {
     public class Statistic {
-        public Statistic (float initialValue, StatType type) {
+        public Statistic (int initialValue, StatType type) {
             Value = initialValue;
             Type = type;
         }
-        public float Value { get; set; }
+        public int Value { get; set; }
         public StatType Type { get; set; }
     }
 }
